+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # MarkdownAST.jl changelog
 
-## Version `v0.1.2`
+## Unreleased
 
+* ![Feature][badge-feature] Implemented `replace` and `replace!` to safely mutate trees in arbitrary ways, and `empty!(node.children)` to remove all the children of a node
 * ![Bugfix][badge-bugfix] The `getproperty` and `setproperty!` methods no longer print unnecessary debug log. ([#19][github-19])
 
 ## Version `v0.1.1`